requirements.txt
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv